requirements.txt
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.38.3
botocore==1.38.3
browser-use==0.1.41
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
Cython==3.0.12
defusedxml==0.7.1
distro==1.9.0
faiss-cpu==1.11.0
filelock==3.18.0
filetype==1.2.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.37
markdownify==1.1.0
MarkupSafe==3.0.2
mem0ai==0.1.88
monotonic==1.6
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
ollama==0.4.8
openai==1.76.0
orjson==3.10.16
packaging==24.2
pillow==11.2.1
playwright==1.51.0
portalocker==2.10.1
posthog==3.25.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
screeninfo==0.8.1
sentence-transformers==4.1.0
setuptools==80.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
zstandard==0.23.0
